server/pyproject.toml
[project]
name = "code-monet"
version = "0.1.0"
description = "Autonomous AI artist that creates ink drawings"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.1.18",
    "pillow>=11.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "svgpathtools>=1.6.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    # Async file I/O
    "aiofiles>=24.1.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
    # AWS
    "boto3>=1.35.0",
    # OpenTelemetry tracing
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "opentelemetry-propagator-aws-xray>=1.0.0",
    "opentelemetry-sdk-extension-aws>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
    "types-aiofiles>=24.1.0",
]

[project.scripts]
code-monet = "code_monet.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["code_monet"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["code_monet"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "boto3-stubs>=1.42.25",
    "types-aiofiles>=25.1.0.20251011",
    "types-python-jose>=3.5.0.20250531",
]
